requirements.txt
psutil==5.9.8
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.23
python-json-logger==2.0.7
python-dotenv
